Rebased OOP modules into separate folder
